requirements.txt
Django==5.1.5
gunicorn==20.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
streamlit==1.42.0
requests==2.32.0
psycopg2-binary==2.9.9
